requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn[standard]
gunicorn
asgiref

# Core Dependencies
python-dotenv
requests
httpx
tenacity

# Data Handling
pandas
simplejson
openpyxl

# Database (if needed for future use)
sqlalchemy
psycopg2-binary==2.9.9
psycopg[pool]==3.2.6   # for connection pooling

# LLM & AI
openai
pydantic==2.11.4
pydantic-ai==0.2.3
pydantic-ai-slim[groq]
pydantic-settings
tiktoken
marqo==3.11.0
rapidfuzz
groq

# Logfire
logfire

# Caching
redis
aiocache

# Authentication
PyJWT
cryptography

# Language and Localization
langcodes
language-data==1.4.0
# Template Engine
jinja2

# AWS
boto3

# Mapbox Geocoding
geopy